+++ src/site/markdown/overview/changelog.md
@@ -19,6 +19,7 @@ This is a bug fixing release.
 
 *   General
     *   [#363](https://github.com/pmd/pmd/issues/363): \[core] Rule documentation pages are missing
+    *   [#364](https://github.com/pmd/pmd/issues/364): \[core] Stream closed exception when running through maven
 
 ### API Changes